requirements.txt
absl-py==1.0.0
affine==2.3.1
aiohttp==3.8.1
aiosignal==1.2.0
analytics-python==1.4.0
anyio==3.6.1
appdirs==1.4.4
asciitree==0.3.3
asgiref==3.5.2
async-timeout==4.0.2
asynctest==0.13.0
attrs==21.4.0
autopep8==1.7.0
backcall==0.2.0
backoff==1.10.0
backports.zoneinfo==0.2.1
bcrypt==3.2.2
beautifulsoup4==4.11.1
black==22.8.0
bleach==5.0.0
bokeh==2.4.3
boltons==21.0.0
boto3==1.24.40
boto3-extensions==0.19.0
botocore==1.27.40
Bottleneck==1.3.5
cached-property==1.5.2
cachetools==4.2.4
certifi==2021.10.8
cffi==1.15.0
cfgrib==0.9.10.3
cfgv==3.3.1
cftime==1.6.2
charset-normalizer==2.0.12
click==8.1.3
click-plugins==1.1.1
click-spinner==0.1.10
cligj==0.7.2
cloudpickle==2.2.0
colorama==0.4.5
colorcet==3.0.1
coverage==6.4
cryptography==37.0.2
cycler==0.11.0
dask==2022.2.0
datashader==0.14.3
datashape==0.5.2
dateparser==1.1.1
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.4
distributed==2022.2.0
docker-pycreds==0.4.0
docopt==0.6.2
eccodes==1.5.0
efficientnet-pytorch==0.7.1
einops==0.6.0
entrypoints==0.4
fastapi==0.78.0
fasteners==0.18
fastjsonschema==2.15.3
ffmpy==0.3.0
filelock==3.7.0
findlibs==0.0.2
fire==0.5.0
fonttools==4.33.3
frozenlist==1.3.0
fsspec==2022.5.0
gantry==0.4.5
gitdb==4.0.9
GitPython==3.1.27
google-auth==2.6.5
google-auth-oauthlib==0.4.6
gradio==3.0.21
grpcio==1.44.0
h11==0.12.0
h5netcdf==1.1.0
h5py==3.6.0
halo==0.0.31
HeapDict==1.0.1
holoviews==1.15.3
httpcore==0.15.0
httpx==0.23.0
hvplot==0.8.2
identify==2.5.1
idna==3.3
importlib-metadata==4.11.3
importlib-resources==5.7.0
iniconfig==1.1.1
ipykernel==4.10.1
ipython==7.32.0
ipython-genutils==0.2.0
ipywidgets==7.7.1
isodate==0.6.1
itermplot==0.331
jedi==0.18.1
Jinja2==2.11.3
jmespath==1.0.1
joblib==1.1.0
jsonschema==4.4.0
jupyter-client==5.3.5
jupyter-core==4.9.2
jupyterlab-widgets==1.1.1
kiwisolver==1.4.2
lightning-utilities==0.3.0
linkify-it-py==1.0.3
llvmlite==0.39.1
locket==1.0.0
log-symbols==0.0.14
Markdown==3.3.6
markdown-it-py==2.1.0
MarkupSafe==2.0.1
marshmallow==3.17.0
marshmallow-oneofschema==3.0.1
matplotlib==3.5.2
matplotlib-inline==0.1.3
mdit-py-plugins==0.3.0
mdurl==0.1.1
mistune==0.8.4
monotonic==1.6
msgpack==1.0.4
multidict==6.0.2
multipledispatch==0.6.0
munch==2.5.0
mypy-extensions==0.4.3
nbconvert==5.6.1
nbformat==5.3.0
nc-time-axis==1.4.1
netCDF4==1.6.2
nltk==3.7
nodeenv==1.6.0
notebook==5.3.1
numba==0.56.4
numbagg==0.2.1
numcodecs==0.10.2
numpy==1.21.6
oauthlib==3.2.0
orjson==3.7.2
packaging==21.3
pandas==1.3.5
pandocfilters==1.5.0
panel==0.14.1
param==1.12.3
paramiko==2.11.0
parso==0.8.3
partd==1.3.0
pathspec==0.10.0
pathtools==0.1.2
pep517==0.13.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.1.2
pip-tools==6.3.1
platformdirs==2.5.2
pluggy==1.0.0
pooch==1.6.0
pre-commit==2.19.0
pretrainedmodels==0.7.4
promise==2.3
prompt-toolkit==3.0.29
protobuf==3.20.0
psutil==5.9.0
ptyprocess==0.7.0
py==1.11.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.9.1
pycparser==2.21
pycryptodome==3.14.1
pyct==0.4.8
pydantic==1.9.1
Pydap==3.2.2
pyDeprecate==0.3.2
pydub==0.25.1
Pygments==2.11.2
PyNaCl==1.5.0
pyparsing==3.0.9
pyproj==3.2.1
pyrsistent==0.18.1
pytest==7.1.1
pytest-cov==3.0.0
python-dateutil==2.8.2
python-multipart==0.0.5
pytorch-lightning==1.8.3
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
pyviz-comms==2.2.1
PyYAML==6.0
pyzmq==22.3.0
rasterio==1.2.10
regex==2022.3.2
requests==2.27.1
requests-oauthlib==1.3.1
rfc3986==1.5.0
rioxarray==0.9.1
rsa==4.8
s3transfer==0.6.0
scipy==1.7.3
seaborn==0.12.1
segmentation-models-pytorch==0.3.1
Send2Trash==1.8.0
sentry-sdk==1.5.9
setproctitle==1.2.2
shortuuid==1.0.8
six==1.16.0
smart-open==5.2.1
smmap==5.0.0
sniffio==1.2.0
snuggs==1.4.7
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
spinners==0.0.24
starlette==0.19.1
tabulate==0.8.10
tblib==1.7.0
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorboardX==2.5.1
termcolor==1.1.0
terminado==0.13.3
testpath==0.6.0
timm==0.4.12
toml==0.10.2
tomli==2.0.1
toolz==0.12.0
torch==1.12.0
torchmetrics==0.11.0
torchvision==0.13.0
tornado==5.1.1
tqdm==4.64.0
traitlets==5.1.1
typed-ast==1.5.4
typeguard==2.13.3
typing-extensions==4.3.0
tzdata==2022.1
tzlocal==4.2
uc-micro-py==1.0.1
urllib3==1.26.9
uvicorn==0.17.6
virtualenv==20.14.1
wandb==0.12.17
wcwidth==0.2.5
webencodings==0.5.1
WebOb==1.8.7
Werkzeug==2.1.2
widgetsnbextension==3.6.1
xarray==0.20.2
xbatcher==0.1.0
yarl==1.7.2
zarr==2.12.0
zict==2.2.0
zipp==3.8.1
